fix(input): NumberInput component moved to Input
