+++ src/form/input/Input.tsx
@@ -2,6 +2,14 @@ import "./_input.scss";
 
 import React from "react";
 import classNames from "classnames";
+import {
+  NOT_INTEGER_FIRST_CHARACTER_OF_STRING_REGEX,
+  NOT_NUMBER_NOR_DECIMAL_POINT_REGEX,
+  PRECISION_REGEX,
+  DECIMAL_NUMBER_SEPARATOR,
+  THOUSANDTHS_SEPARATOR,
+  IS_LAST_CHARACTER_DECIMAL_POINT_REGEX
+} from "../../core/utils/number/numberConstants";
 
 type InputTypes =
   | "checkbox"
@@ -40,6 +48,7 @@ export type InputProps = Omit<
   hasError?: boolean;
   customClassName?: string;
   inputContainerRef?: React.RefObject<HTMLDivElement>;
+  maxFractionDigits?: number;
 };
 
 function Input(props: InputProps) {
@@ -55,6 +64,8 @@ function Input(props: InputProps) {
     autoComplete = "off",
     autoCorrect = "off",
     inputContainerRef,
+    onChange,
+    maxFractionDigits = 0,
     ...rest
   } = props;
   const inputContainerClassName = classNames("input-container", customClassName);
@@ -62,6 +73,7 @@ function Input(props: InputProps) {
     "input--is-disabled": isDisabled,
     "input--has-error": hasError
   });
+  const isNumberInput = type === "number";
 
   return (
     <div
@@ -77,10 +89,11 @@ function Input(props: InputProps) {
 
       <input
         className={inputClassName}
-        type={type}
+        type={isNumberInput ? "text" : type}
         autoComplete={autoComplete}
         autoCorrect={autoCorrect}
         disabled={isDisabled}
+        onChange={isNumberInput ? handleChange : onChange}
         {...rest}
       />
 
@@ -91,6 +104,36 @@ function Input(props: InputProps) {
       )}
     </div>
   );
+
+  function handleChange(event: React.SyntheticEvent<HTMLInputElement>) {
+    const {value: newValue} = event.currentTarget;
+
+    let formattedNewValue = newValue
+      .replace(NOT_INTEGER_FIRST_CHARACTER_OF_STRING_REGEX, "")
+      .replace(NOT_NUMBER_NOR_DECIMAL_POINT_REGEX, "")
+      .replace(PRECISION_REGEX, "$1");
+
+    if (maxFractionDigits > 0) {
+      const decimalNumberParts = newValue.split(DECIMAL_NUMBER_SEPARATOR);
+      const decimalPart = decimalNumberParts[1];
+      const integerPart = decimalNumberParts[0].replace(THOUSANDTHS_SEPARATOR, "");
+
+      if (decimalPart && decimalPart.length > maxFractionDigits) {
+        const trimmedDecimalPart = decimalPart.slice(0, maxFractionDigits);
+
+        formattedNewValue = `${integerPart}${DECIMAL_NUMBER_SEPARATOR}${trimmedDecimalPart}`;
+      }
+    } else {
+      formattedNewValue = formattedNewValue.replace(
+        IS_LAST_CHARACTER_DECIMAL_POINT_REGEX,
+        ""
+      );
+    }
+
+    if (formattedNewValue !== newValue) {
+      event.currentTarget.value = formattedNewValue;
+    }
+  }
 }
 
 export default Input;
